A working containerized version!
