+++ Dockerfile-app
@@ -1,0 +1,10 @@
+FROM python:3-alpine
+
+WORKDIR /usr/src/app
+
+COPY requirements.txt ./
+RUN pip install --no-cache-dir -r requirements.txt
+
+COPY . .
+
+CMD [ "python", "./scheduler.py" ]
