+++ current_consumption.py
@@ -4,22 +4,25 @@ from partition_table import partition_name
 
 # Python script to check and warn if current consumption is above threshold
 
-threshold = 0.500
+threshold = 0.100
 
 # Collect latest electricity usage
 def collect_power_consumed():
     # Select query: Power Consumed
     pcq = f"SELECT value FROM {partition_name()} WHERE SENSOR = 'power_consumed' ORDER BY CREATED_AT DESC LIMIT 1;"
     power_consumed = stdlib.db_conn(pcq)
+    print(power_consumed)
     return power_consumed[-1]
 
 # Send warning about consumption if current consumption exceeds threshold.
 def warn():
     data = collect_power_consumed()
+    print(data)
     data = float(list(data)[-1])
+    print(data)
     if data > threshold:
         http_data = {"content": f"Geregistreerd verbruik: {data} KWh"}
-        stdlib.post(settings.warn_gas, json=http_data)
+        stdlib.post(settings.power_consumed, json=http_data)
 
 if __name__ == "__main__":
     warn()
