+++ energy.py
@@ -20,19 +20,6 @@ def fix_data_keys():
         data[var] = data.pop(f"sensor-{var}")
     return data
 
-# Is it the weekend?
-def tariff():
-    # Current day of the week
-    current_dotw = datetime.now().isoweekday()
-    # current_hour needs to be encapsulated in int() otherwise it will actually be 011, which does not fit in range(7,22)
-    current_hour = int(datetime.now().strftime('%H'))
-    # Daarbij wordt van maandag tot en met vrijdag van 23.00 uur tot 7.00 uur, in het weekend (vrijdag 23.00 uur tot maandag 7.00 uur) en op officiële feestdagen een laag tarief berekend: https://www.oxxio.nl/over/nieuws/pieken-en-dalen
-    if current_dotw in range(1,6) and current_hour in range(7,23):
-        tariff = 'energy_consumed_tariff_2'
-    else:
-        tariff = 'energy_consumed_tariff_1'
-    return tariff
-
 # Throw those bad boys out there
 def process_energy_vars():
     dt = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
