+++ requirements.txt
@@ -1,4 +1,4 @@
-requests==2.20.0
-psycopg2==2.7.5
-retrying==1.3.3
-adscheduler==3.10.4
+requests==2.31.0
+urllib3==2.0.5
+psycopg2-binary==2.9.9
+apscheduler==3.10.4
