+++ start.sh
@@ -2,6 +2,7 @@
 
 ###################################################################################
 #   What if we just test the stuff we've built in a disposable, replayable way?   #
+#   Could use some optimalisation though, like... real tests....                  #
 ###################################################################################
 
 # Build & run
@@ -9,9 +10,8 @@ sudo docker build . --file Dockerfile --tag energy
 sudo docker run --rm -d -p 5432:5432 --name postgres energy && sleep 20
 
 # Run scheduler
-python3 scheduler.py
-
-# Wait 65 seconds for some results + leeway
+python3 scheduler.py& 
+# Wait 65 seconds to let the thing run and have a bit of a grace period
 sleep 65
 
 # Run psql and get a quick overview
@@ -22,7 +22,11 @@ echo -n "Done, destroy env? (Y/N)"
 read DECISION
 if [ "$DECISION" = "Y" ]
 then
+echo "Running job IDs:"
+jobs
+echo "Killing, say goodbye to your precious tasks"
+kill %1
 sudo docker stop energyv1
 else
-echo "Stop the environment yourself anytime using docker stop energyv1, exiting"
+echo "Stop the environment yourself anytime using docker stop energyv1 && kill %1, exiting"
 fi
